+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ### vNEXT
 - Allow passing of a minimum connection timeout for connect/reconnect
 
+### v0.9.16
+- Add ability to set custom WebSocket protocols for client. <br/>
+  [@pkosiec](https://github.com/pkosiec) in [#477](https://github.com/apollographql/subscriptions-transport-ws/pull/477)
+
+### v0.9.15
+
+- Add support for `graphql` and `@types/graphql` 14.  <br/>
+  [@caiquecastro](https://github.com/caiquecastro) in [#464](https://github.com/apollographql/subscriptions-transport-ws/pull/464)
+
 ### v0.9.14
 - Allow dynamically specifying/overriding the schema in the object returned from `onOperation` [PR #447](https://github.com/apollographql/subscriptions-transport-ws/pull/447)